Convert all floats to doubles, vector2f to vector2d

+++ src/rrt/2dplane/2dplane.cpp
@@ -7,10 +7,10 @@ using namespace Eigen;
 using namespace RRT;
 using namespace std;
 
-Tree<Vector2f>* RRT::TreeFor2dPlane(
-    shared_ptr<StateSpace<Eigen::Vector2f>> stateSpace, Vector2f goal,
-    float step) {
-    Tree<Vector2f>* rrt = new Tree<Vector2f>(stateSpace);
+Tree<Vector2d>* RRT::TreeFor2dPlane(
+    shared_ptr<StateSpace<Eigen::Vector2d>> stateSpace, Vector2d goal,
+    double step) {
+    Tree<Vector2d>* rrt = new Tree<Vector2d>(stateSpace);
 
     rrt->setStepSize(step);
